Fix the server hang in case of long prompts/requests in certain serve… (#89)
* Fix the server hang in case of long prompts/requests in certain server configuration

There is a discrepancy between limitations of the model, user and cache.
To avoid infinite loop we need to discard long requests exceeding
server limitations and/or handle long requests more agile

* remove unnecessary verification in add request

* Refactor validation conditions during request add

* Make one more condition simplification

* Remove configuration verification bypass and fix tests

* Fix conditins to discard requests from inference

* Add test on sync engine hang for two use cases

* Add staged engine tests

* Move DummaryModel from tests to mlc_serve

* Rename non pytest test functions to exclude from regualar pytest testing

* address PR comments

* Clean-up prompt_allocate_ratio/prompt-allocat-_ratio

* Rename unittest files

* Beautify comment

* Add test for request exceed cache limit, fix comment

* fix formatting

* Add comment

* Comments beautification

* Change Dummary nameing to Dummy
